requirements.txt
asarPy==1.0.1
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
stripe==12.4.0
typing_extensions==4.14.1
urllib3==2.5.0
